Change script for running tests
A new script is created to run tests that does not specify the testsuite
from the Open Social phpunit configuration. The script also allows
passing extra flags to PHPUnit which makes it easy to select specific
classes or testsuites without having to figure out the specific command
for PHPUnit using Open Social's test configuration.

The existing unit-test.sh file is kept and targets specifically the
social test-suite so that Travis keeps working.