Fix local script calling
The working directory when Travis executes the `unit-test.sh` script is
not the script directory, this means a relative path to call another
script doesn't work.

The files are also given executable permissions so that Travis no longer
needs to manually invoke a shell to execute the script. This ensures the
variable substitution needed to find the script's directory works (that
pattern isn't supported by sh or dash).